// src/app/Models/InterviewSchedule.ts
export class interviewSchedule {
  candidateId: string;
  candidateName: string;
  createdBy: string;
  createdByName: string;
  createdDateTime: string;
  duration: string;
  id: string;
  interviewerId: string;
  interviewerName: string;
  jrNumber: string;
  message: string;
  recStatus: string;
  resultRemarks: string;
  round: string;
  scheduleDateTime: string;
  scheduleRemarks: string;
  status: string;
  updatedBy: string;
  updatedByName: string;
  updatedDateTime: string;
  mode: string;
  venue: string;
  timeZone: string;
  startsFrom: string;
  fromDateTime: string;
  toDateTime: string;
}